Fix Verus HumanEval compilation issues and expand dataset
- Fixed 4 non-compiling files from special folders:
  * humaneval_103.rs: Added decreases clause to recursive function
  * humaneval_047.rs: Fixed type mismatch and spec function issues
  * humaneval_007.rs: Fixed trigger inference with recursive approach
  * humaneval_154_cycpattern_check.rs: Fixed trigger inference issues
- Moved some specs outside helpers
- Integrated 2 files from poor/default_values:
  * humaneval_140.rs: Space transformation with triggers
  * humaneval_146.rs: Special filter with complete specification

- Updated all corresponding YAML files with proper specifications
- Fixed map function signatures across multiple files (|_i: int, x: T|)
- Resolved cast expression syntax issues (parentheses around casts)
- Improved trigger annotations for better SMT solver performance

Dataset expansion: 155 → 161 working file pairs (+3.9% increase)
All files now compile successfully with Verus

+++ benchmarks/verus/humaneval/files/humaneval_068.rs
@@ -55,7 +55,7 @@ fn pluck(arr: Vec<i8>) -> (result: Vec<i8>)
         arr@.len() == 0 ==> result@.len() == 0,
         !has_even_value(arr@.map_values(|x: i8| x as int)) ==> result@.len() == 0,
         has_even_value(arr@.map_values(|x: i8| x as int)) ==> result@.len() == 2,
-        result@.len() == 2 ==> (0 <= result@[1] as int && result@[1] as int < arr@.len()),
+        result@.len() == 2 ==> (0 <= (result@[1] as int) && (result@[1] as int) < (arr@.len() as int)),
         result@.len() == 2 ==> arr@[result@[1] as int] as int == result@[0] as int,
         result@.len() == 2 ==> result@[0] as int % 2 == 0,
         result@.len() == 2 ==> forall|i: int| 0 <= i < arr@.len() && arr@[i] as int % 2 == 0 ==> result@[0] as int <= arr@[i] as int,
